ordershield-network/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
web3==6.15.1
python-dotenv==1.0.0
streamlit==1.30.0
requests==2.31.0
sqlalchemy==2.0.25
